+++ packages/machines/floating-panel/src/floating-panel.connect.ts
@@ -61,6 +61,7 @@ export function connect<T extends PropTypes>(
     getTriggerProps() {
       return normalize.button({
         ...parts.trigger.attrs,
+        dir: prop("dir"),
         type: "button",
         disabled: prop("disabled"),
         id: dom.getTriggerId(scope),
@@ -79,6 +80,7 @@ export function connect<T extends PropTypes>(
     getPositionerProps() {
       return normalize.element({
         ...parts.positioner.attrs,
+        dir: prop("dir"),
         id: dom.getPositionerId(scope),
         style: {
           "--width": toPx(size?.width),
@@ -95,6 +97,7 @@ export function connect<T extends PropTypes>(
     getContentProps() {
       return normalize.element({
         ...parts.content.attrs,
+        dir: prop("dir"),
         role: "dialog",
         tabIndex: 0,
         hidden: !open,
@@ -152,6 +155,7 @@ export function connect<T extends PropTypes>(
     getCloseTriggerProps() {
       return normalize.button({
         ...parts.closeTrigger.attrs,
+        dir: prop("dir"),
         disabled: prop("disabled"),
         "aria-label": "Close Window",
         type: "button",
@@ -186,11 +190,13 @@ export function connect<T extends PropTypes>(
 
       return normalize.button({
         ...parts.stageTrigger.attrs,
+        dir: prop("dir"),
         disabled: prop("disabled"),
         ...actionProps,
         type: "button",
         onClick(event) {
           if (event.defaultPrevented) return
+          if (!prop("resizable")) return
           const type = match(props.stage, {
             minimized: () => "MINIMIZE",
             maximized: () => "MAXIMIZE",
@@ -204,6 +210,7 @@ export function connect<T extends PropTypes>(
     getResizeTriggerProps(props: ResizeTriggerProps) {
       return normalize.element({
         ...parts.resizeTrigger.attrs,
+        dir: prop("dir"),
         "data-disabled": dataAttr(!canResize),
         "data-axis": props.axis,
         onPointerDown(event) {
@@ -237,6 +244,7 @@ export function connect<T extends PropTypes>(
     getDragTriggerProps() {
       return normalize.element({
         ...parts.dragTrigger.attrs,
+        dir: prop("dir"),
         "data-disabled": dataAttr(!canDrag),
         onPointerDown(event) {
           if (!canDrag) return
@@ -263,7 +271,9 @@ export function connect<T extends PropTypes>(
             node.releasePointerCapture(event.pointerId)
           }
         },
-        onDoubleClick() {
+        onDoubleClick(event) {
+          if (event.defaultPrevented) return
+          if (!prop("resizable")) return
           send({ type: isMaximized ? "RESTORE" : "MAXIMIZE" })
         },
         style: {
@@ -278,6 +288,7 @@ export function connect<T extends PropTypes>(
     getControlProps() {
       return normalize.element({
         ...parts.control.attrs,
+        dir: prop("dir"),
         "data-disabled": dataAttr(prop("disabled")),
         "data-stage": context.get("stage"),
         "data-minimized": dataAttr(isMinimized),
@@ -289,6 +300,7 @@ export function connect<T extends PropTypes>(
     getTitleProps() {
       return normalize.element({
         ...parts.title.attrs,
+        dir: prop("dir"),
         id: dom.getTitleId(scope),
       })
     },
@@ -296,6 +308,7 @@ export function connect<T extends PropTypes>(
     getHeaderProps() {
       return normalize.element({
         ...parts.header.attrs,
+        dir: prop("dir"),
         id: dom.getHeaderId(scope),
         "data-dragging": dataAttr(dragging),
         "data-topmost": dataAttr(isTopmost),
@@ -309,6 +322,7 @@ export function connect<T extends PropTypes>(
     getBodyProps() {
       return normalize.element({
         ...parts.body.attrs,
+        dir: prop("dir"),
         "data-dragging": dataAttr(dragging),
         "data-minimized": dataAttr(isMinimized),
         "data-maximized": dataAttr(isMaximized),