+++ packages/machines/navigation-menu/src/index.ts
@@ -8,9 +8,11 @@ export type {
   NavigationMenuProps as Props,
   NavigationMenuService as Service,
   ArrowProps,
+  ContentProps,
   ElementIds,
   ViewportProps,
   ItemProps,
+  ItemState,
   LinkProps,
   ValueChangeDetails,
 } from "./navigation-menu.types"