+++ packages/machines/navigation-menu/src/navigation-menu.types.ts
@@ -129,9 +129,51 @@ export interface ItemProps {
   disabled?: boolean | undefined
 }
 
+export interface ItemState {
+  /**
+   * The id of the item element
+   */
+  itemId: string
+  /**
+   * The id of the trigger element
+   */
+  triggerId: string
+  /**
+   * The id of the trigger proxy element
+   */
+  triggerProxyId: string
+  /**
+   * The id of the content element
+   */
+  contentId: string
+  /**
+   * Whether the item is currently selected
+   */
+  selected: boolean
+  /**
+   * Whether the item was previously selected (for animation)
+   */
+  wasSelected: boolean
+  /**
+   * Whether the item's content is open
+   */
+  open: boolean
+  /**
+   * Whether the item is disabled
+   */
+  disabled: boolean
+}
+
 export interface ArrowProps {
   /**
    * The value of the item
+   */
+  value: string
+}
+
+export interface ContentProps {
+  /**
+   * The value of the item this content belongs to
    */
   value: string
 }
@@ -149,6 +191,11 @@ export interface LinkProps {
    * Function called when the link is selected
    */
   onSelect?: ((event: CustomEvent) => void) | undefined
+  /**
+   * Whether to close the navigation menu when the link is clicked.
+   * @default true
+   */
+  closeOnClick?: boolean | undefined
 }
 
 export interface ViewportProps {
@@ -184,12 +231,16 @@ export interface NavigationMenuApi<T extends PropTypes = PropTypes> {
    * The orientation of the menu
    */
   orientation: Orientation
+  /**
+   * Function to reposition the viewport
+   */
+  reposition: VoidFunction
 
   getRootProps: () => T["element"]
   getListProps: () => T["element"]
   getItemProps: (props: ItemProps) => T["element"]
-  getIndicatorTrackProps: () => T["element"]
   getIndicatorProps: () => T["element"]
+  getItemIndicatorProps: (props: ItemProps) => T["element"]
   getArrowProps: (props?: ArrowProps) => T["element"]
 
   getTriggerProps: (props: ItemProps) => T["button"]
@@ -197,7 +248,9 @@ export interface NavigationMenuApi<T extends PropTypes = PropTypes> {
   getViewportProxyProps: (props: ItemProps) => T["element"]
 
   getLinkProps: (props: LinkProps) => T["element"]
-  getContentProps: (props: LinkProps) => T["element"]
+  getContentProps: (props: ContentProps) => T["element"]
   getViewportPositionerProps: (props?: ViewportProps) => T["element"]
   getViewportProps: (props?: ViewportProps) => T["element"]
+
+  getItemState: (props: ItemProps) => ItemState
 }