+++ packages/machines/popover/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @zag-js/popover
 
+## 1.30.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.30.0
+  - @zag-js/core@1.30.0
+  - @zag-js/types@1.30.0
+  - @zag-js/aria-hidden@1.30.0
+  - @zag-js/utils@1.30.0
+  - @zag-js/dismissable@1.30.0
+  - @zag-js/dom-query@1.30.0
+  - @zag-js/focus-trap@1.30.0
+  - @zag-js/popper@1.30.0
+  - @zag-js/remove-scroll@1.30.0
+
+## 1.29.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/anatomy@1.29.1
+  - @zag-js/core@1.29.1
+  - @zag-js/types@1.29.1
+  - @zag-js/aria-hidden@1.29.1
+  - @zag-js/utils@1.29.1
+  - @zag-js/dismissable@1.29.1
+  - @zag-js/dom-query@1.29.1
+  - @zag-js/focus-trap@1.29.1
+  - @zag-js/popper@1.29.1
+  - @zag-js/remove-scroll@1.29.1
+
 ## 1.29.0
 
 ### Patch Changes