+++ packages/utilities/dom-query/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @zag-js/dom-query
 
+## 1.30.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/types@1.30.0
+
+## 1.29.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zag-js/types@1.29.1
+
 ## 1.29.0
 
 ### Patch Changes